convert-pdf-to-doc/requirements.txt
fire==0.5.0
fonttools==4.40.0
# lxml will not install on my arm M1 mac, instead installing
# via Dockerfile OS command
#lxml==4.9.3
numpy==1.25.0
opencv-python==4.8.0.74
pdf2docx==0.5.6
pip==21.2.3
PyMuPDF==1.22.5
python-docx==0.8.11
python-dotenv==1.0.0
setuptools==57.4.0
six==1.16.0
termcolor==2.3.0

